requirements.txt
ollama
streamlit
gradio
pdfplumber
langchain
langchain-core
langchain-ollama
langchain_community
langchain_text_splitters
unstructured
unstructured[all-docs]
onnx==1.16.1
protobuf==3.20.3
chromadb==0.4.22
Pillow
numpy
st-gsheets-connection
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
plotly